feat (eda) Initial Exploratory Analysis
Examine the raw data set identify any duplicate values or data type
which will have to be corrected for analysis. Analyze the number of
apps in each category to better understand which categories are most
common on the marketplace. Plot distribution of the app ratings to
identify that most apps are highly rated with few exceptions. Plot
the counts of Paid vs Free apps identifying free apps are the most
common across all categories.